fixed orders

+++ shopping_cart/models.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import User
 from home.models import Movie
 class Order(models.Model):
     id = models.AutoField(primary_key=True)
-    total = models.IntegerField()
+    total = models.DecimalField(max_digits=10, decimal_places=2)
     date = models.DateTimeField(auto_now_add=True)
     user = models.ForeignKey(User,
         on_delete=models.CASCADE)
